move mainloop into separate file
